Separa clases y funciones en sus propios archivos
