+++ src/utils.ts
@@ -1,0 +1,3 @@
+export function randomIntFromInterval(min,max) {// min and max included
+    return Math.floor(Math.random()*(max-min+1)+min);
+}
